+++ normalize/exc.py
@@ -147,6 +147,13 @@ class CollectionCoerceError(CoercionError):
     message = "Cannot interpret {giventype} as a {fortype} constructor"
 
 
+class CollectionItemCoerceError(CoercionError):
+    message = (
+        "coerce to {valuetype} for insertion to {colltype} failed for "
+        "value {value}: {exc}"
+    )
+
+
 class CollectionDefinitionError(SubclassError):
     message = "{property} must be defined in a {coll} subclass"
 
@@ -175,6 +182,20 @@ class DefaultSignatureError(PropertyDefinitionError):
     message = (
         "default functions must have no required arguments (except "
         "'self'); {module}.{func} wants {nargs}"
+    )
+
+
+class DictOfWhat(PropertyDefinitionError):
+    message = (
+        "Dict Properties must have a defined item type; pass of= "
+        "or dict_of= to the declaration"
+    )
+
+
+class DictPropertyMustDeriveDictCollection(PropertyDefinitionError):
+    message = (
+        "Dict Property collections must derive DictCollection, and "
+        "{got} doesn't"
     )